services/optimizer/gepa_runner/requirements.txt
dspy-ai>=2.5,<3
langfuse>=2.0,<3
apscheduler>=3.10,<4
psycopg[binary]>=3.2,<4
fastapi>=0.111,<0.120
uvicorn[standard]>=0.30,<0.35
httpx>=0.27,<0.30
litellm>=1.50,<2
